add unit test add image and doc fix

+++ java/src/main/java/com/tizzy/business/repository/DataManager.java
@@ -2,33 +2,26 @@ package com.tizzy.business.repository;
 
 import org.apache.groovy.util.concurrent.concurrentlinkedhashmap.ConcurrentLinkedHashMap;
 import org.apache.groovy.util.concurrent.concurrentlinkedhashmap.Weighers;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class DataManager {
+
+    private static final Logger logger = LoggerFactory.getLogger(DataManager.class);
 
     private final static long dataLimit = 1000000L;
     private final static long cacheLimit = 10000L;
     private static DataManager instance = null;
 
-    //短码code到url长链的数据表
-    private ConcurrentLinkedHashMap<Long, Url> memDB;
-//    = new ConcurrentLinkedHashMap.Builder<Long, Url>()
-//            .maximumWeightedCapacity(dataLimit).weigher(Weighers.singleton())
-//            .build();
-
-    //长链url到短码code的缓存
-    private ConcurrentLinkedHashMap<String, Long> urlCache;
-//    = new ConcurrentLinkedHashMap.Builder<String, Long>()
-//            .maximumWeightedCapacity(cacheLimit).weigher(Weighers.singleton())
-//            .build();
+    private static ConcurrentLinkedHashMap<Long, Url> memDB = new ConcurrentLinkedHashMap.Builder<Long, Url>()
+            .maximumWeightedCapacity(dataLimit).weigher(Weighers.singleton())
+            .build();
+    private ConcurrentLinkedHashMap<String, Long> urlCache = new ConcurrentLinkedHashMap.Builder<String, Long>()
+            .maximumWeightedCapacity(cacheLimit).weigher(Weighers.singleton())
+            .build();
 
     private DataManager() {
-        memDB = new ConcurrentLinkedHashMap.Builder<Long, Url>()
-                .maximumWeightedCapacity(dataLimit).weigher(Weighers.singleton())
-                .build();
 
-        urlCache = new ConcurrentLinkedHashMap.Builder<String, Long>()
-                .maximumWeightedCapacity(cacheLimit).weigher(Weighers.singleton())
-                .build();
     }
 
     public static DataManager getInstance() {
@@ -55,7 +48,7 @@ public class DataManager {
             Url urlData = memDB.get(urlCode);
 
             //数据过期，手动移除
-            if (System.currentTimeMillis() >= urlData.getExpiresDate()) {
+            if (urlData != null && System.currentTimeMillis() >= urlData.getExpiresDate()) {
                 memDB.remove(urlCode);
                 urlCache.remove(longUrl);
                 result = 0;
@@ -67,6 +60,8 @@ public class DataManager {
     public void put(long urlCode, String longUrl, long expiresTime) {
         memDB.put(urlCode, new Url(longUrl, expiresTime));
         urlCache.put(longUrl, urlCode);
+
+        logger.info("put data urlCode {}, longUrl {}", urlCode, longUrl);
     }
 
     public String get(long urlCode) {
